+++ src/main/kotlin/no/nav/familie/ef/sak/repository/BehandlingHistorikkRepository.kt
@@ -1,10 +1,12 @@
 package no.nav.familie.ef.sak.repository
 
-import no.nav.familie.ef.sak.repository.domain.BehandlingsHistorikk
+import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
 import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface BehandlingHistorikkRepository : RepositoryInterface<BehandlingsHistorikk, UUID>, InsertUpdateRepository<BehandlingsHistorikk>{
-    fun findByBehandlingId(behandlingId: UUID): List<BehandlingsHistorikk>
+interface BehandlingHistorikkRepository : RepositoryInterface<BehandlingHistorikk, UUID>,
+                                          InsertUpdateRepository<BehandlingHistorikk> {
+
+    fun findByBehandlingId(behandlingId: UUID): List<BehandlingHistorikk>
 }
