--- src/main/kotlin/no/nav/familie/ef/sak/repository/domain/BehandlingsHistorikk.kt
@@ -1,14 +1,0 @@
-package no.nav.familie.ef.sak.repository.domain
-
-import no.nav.familie.ef.sak.service.steg.StegType
-import org.springframework.data.annotation.Id
-import java.time.LocalDateTime
-import java.util.*
-
-data class BehandlingsHistorikk(@Id
-                                val id: UUID = UUID.randomUUID(),
-                                val behandlingId: UUID,
-                                var steg: StegType,
-                                val endretAvNavn: String,
-                                val endretAvMail: String,
-                                val endretTid: LocalDateTime = LocalDateTime.now())
