+++ src/main/kotlin/no/nav/familie/ef/sak/service/BehandlingHistorikkService.kt
@@ -1,20 +1,14 @@
 package no.nav.familie.ef.sak.service
 
 import no.nav.familie.ef.sak.repository.BehandlingHistorikkRepository
-import no.nav.familie.ef.sak.repository.domain.BehandlingsHistorikk
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
+import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 import java.util.*
 
 @Service
 class BehandlingHistorikkService(val behandlingHistorikkRepository: BehandlingHistorikkRepository) {
 
-    val logger: Logger = LoggerFactory.getLogger(this.javaClass)
-
-    @Transactional
-    fun finnBehandling(behandlingId: UUID) : List<BehandlingsHistorikk> {
+    fun finnBehandlingHistorikk(behandlingId: UUID) : List<BehandlingHistorikk> {
         return behandlingHistorikkRepository.findByBehandlingId(behandlingId)
     }
